iox-#14 add tests for ChunkSettings and remove MemoryManager::requiredChunkSize
