+++ iceoryx_posh/include/iceoryx_posh/internal/mepoo/memory_manager.hpp
@@ -63,19 +63,6 @@ class MemoryManager
 
     MemPoolInfo getMemPoolInfo(uint32_t f_index) const noexcept;
 
-    /// @brief calculates the required chunk size regarding the constraints on payload and custom header
-    /// @param[in] payloadSize is the size of the user payload without additional headers
-    /// @param[in] payloadAlignment is the alignment of the user payload
-    /// @param[in] customHeaderSize is the size of the custom user header; use iox::CHUNK_NO_CUSTOM_HEADER_SIZE to omit
-    /// a custom header
-    /// @param[in] customHeaderAlignment is the alignment of the custom user header; use
-    /// iox::CHUNK_NO_CUSTOM_HEADER_ALIGNMENT to omit a custom header
-    /// @return the minimum chunk size required to hold the user payload with the custom header
-    static uint32_t requiredChunkSize(const uint32_t payloadSize,
-                                      const uint32_t payloadAlignment,
-                                      const uint32_t customHeaderSize,
-                                      const uint32_t customHeaderAlignment) noexcept;
-
     static uint64_t requiredChunkMemorySize(const MePooConfig& f_mePooConfig) noexcept;
     static uint64_t requiredManagementMemorySize(const MePooConfig& f_mePooConfig) noexcept;
     static uint64_t requiredFullMemorySize(const MePooConfig& f_mePooConfig) noexcept;
